requirements.txt
fastapi==0.115.0  
uvicorn==0.32.0
openai>=1.54.0
pydantic==2.10.0
requests==2.32.3
httpx==0.28.0
python-dotenv==1.0.1

